Bring Files More In Line With Other Create Funcs

+++ Engine/src/Oyl3D/Platform/Windows/Win32Platform.cpp
@@ -5,7 +5,7 @@ namespace oyl {
 
 Platform* Platform::s_instance = new Win32Platform;
 
-double Win32Platform::getTimeImpl() {
+float Win32Platform::getTimeImpl() {
 	LARGE_INTEGER li;
 	static uint64_t timeStart = [&]()->float {
 		QueryPerformanceCounter(&li);
