+++ Engine/src/Oyl3D/System/Platform.h
@@ -5,9 +5,9 @@ namespace oyl {
 class Platform {
 public:
 	// Returns current time since execution in seconds
-	static double getTime() { return s_instance->getTimeImpl(); };
+	static float getTime() { return s_instance->getTimeImpl(); };
 private:
-	virtual double getTimeImpl() = 0;
+	virtual float getTimeImpl() = 0;
 private:
 	static Platform* s_instance;
 };
